email_service/email_service/crons/requirements.txt
Django
djangorestframework
pulsar-client==3.4.0
